refactor: extract getNavigationList function
