+++ homepage/src/lib/getNavigationList.js
@@ -26,3 +26,23 @@ export async function getPagesList(lang) {
 
   return Promise.all(pagesList);
 }
+
+export async function getNavigationList(lang) {
+  const pagesList = await getPagesList(lang);
+  const cardsPage = {
+    en: 'Cards',
+    'zh-tw': '卡片頁',
+  };
+
+  // * dynamic page navigation
+  const navigationList = pagesList
+    .filter((page) => page.path && page.name)
+    .map((page) => {
+      page.path = page.path.replace('index', '');
+      return { link: `/${page.path}`, text: page.name };
+    });
+  // add hard coded cards page
+  navigationList.push({ link: `/cards`, text: cardsPage[lang] });
+
+  return navigationList;
+}
